# apispec_1.json
{
...
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "A swagger API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/citoyen/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Citoyen cr\u00e9\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes pour la cr\u00e9ation de citoyen"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Ajoute un nouveau citoyen."
      }
    },
    "/api/citoyen/all": {
      "get": {
        "responses": {
          "200": {
            "description": "Liste des citoyens"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les citoyens enregistr\u00e9s."
      }
    },
    "/api/citoyen/delete/{citoyen_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "citoyen_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "Citoyen non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Supprime un citoyen de la base de donn\u00e9es."
      }
    },
    "/api/citoyen/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connexion r\u00e9ussie"
          },
          "400": {
            "description": "Identifiants manquants"
          },
          "401": {
            "description": "Identifiants invalides"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Connecte un citoyen."
      }
    },
    "/api/citoyen/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "D\u00e9connexion r\u00e9ussie"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "D\u00e9connecte un citoyen."
      }
    },
    "/api/citoyen/protected": {
      "get": {
        "responses": {
          "200": {
            "description": "D\u00e9tails du citoyen connect\u00e9"
          },
          "404": {
            "description": "Citoyen non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re les informations du citoyen connect\u00e9."
      }
    },
    "/api/citoyen/register": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Enregistrement r\u00e9ussi"
          },
          "400": {
            "description": "Champs obligatoires manquants ou type utilisateur invalide"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Inscrit un nouveau citoyen."
      }
    },
    "/api/citoyen/update/{citoyen_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Citoyen mis \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es manquantes"
          },
          "404": {
            "description": "Citoyen non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Met \u00e0 jour les informations d'un citoyen."
      }
    },
    "/api/citoyen/{citoyen_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "citoyen_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails du citoyen"
          },
          "404": {
            "description": "Citoyen non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re un citoyen sp\u00e9cifique via son ID."
      }
    },
...
    "/api/user/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "user_type": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Utilisateur cr\u00e9\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes pour cr\u00e9er un utilisateur"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Ajoute un nouvel utilisateur."
      }
    },
    "/api/user/all": {
      "get": {
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les utilisateurs enregistr\u00e9s."
      }
    },
    "/api/user/delete/{user_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Supprime un utilisateur de la base de donn\u00e9es."
      }
    },
    "/api/user/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connexion r\u00e9ussie"
          },
          "400": {
            "description": "Identifiants requis"
          },
          "401": {
            "description": "Authentification \u00e9chou\u00e9e"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Connecte un utilisateur."
      }
    },
    "/api/user/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "D\u00e9connexion r\u00e9ussie"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "D\u00e9connecte un utilisateur."
      }
    },
    "/api/user/protected": {
      "get": {
        "responses": {
          "200": {
            "description": "D\u00e9tails de l'utilisateur connect\u00e9"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re les informations de l'utilisateur connect\u00e9."
      }
    },
    "/api/user/register": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "user_type": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Enregistrement r\u00e9ussi"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes ou type d'utilisateur invalide"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Inscrit un nouvel utilisateur."
      }
    },
    "/api/user/update/{user_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur mis \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es manquantes"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Met \u00e0 jour les informations d'un utilisateur."
      }
    },
    "/api/user/{user_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails de l'utilisateur"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re un utilisateur sp\u00e9cifique via son ID."
      }
    },
...
  },
  "swagger": "2.0"
}